pyproject.toml
[project]
name = "darkir"
version = "0.1.0"
description = "DarkIR macOS ARM setup"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.123.9",
    "pillow>=12.0.0",
    "ptflops>=0.7.5",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]
